Refactoring widget to be a custom element.
