+++ src/components/feedback-widgets/feedback-widget-custom-element.shadow.css
@@ -1,0 +1,442 @@
+:host {
+	& slot[name="title"] {
+		@media (width < 1024px) {
+			display: none;
+		}
+	}
+
+	& .widget_content {
+		/* Layout */
+		position: relative;
+
+		& h3 {
+			/* Layout */
+			margin-block-end: 5--step;
+
+			/* Text */
+			font-size: 20--rpx;
+			font-weight: 500;
+		}
+
+		& p {
+			/* Text */
+			font-size: 16--rpx;
+		}
+
+		& button {
+			/* Layout */
+			margin-block-start: 6--step;
+
+			/* Layout */
+			padding-block: 1--step;
+			padding-inline: 4--step;
+
+			/* Text */
+			font-size: 16--rpx;
+			line-height: 24--rpx;
+
+			/* Appearance */
+			background-color: #ffffff;
+
+			/* Appearance */
+			border: 1px solid #005a8f;
+
+			/* Appearance */
+			border-radius: 3px;
+			cursor: pointer;
+
+			&.widget_vote-button {
+				/* Layout */
+				padding: 2--step;
+
+				&:first-child {
+					/* Layout */
+					margin-inline-end: 2--step;
+				}
+
+				& svg {
+					/* Text */
+					vertical-align: middle;
+
+					& path {
+						fill: #005a8f;
+					}
+				}
+
+				&.selected {
+					/* Appearance */
+					background-color: #2f7aa7;
+					border: 1px solid #005a8f;
+
+					& svg path {
+						fill: #ffffff;
+					}
+				}
+			}
+
+			&.widget_primary-button {
+				/* Appearance */
+				background-color: #005a8f;
+				color: #ffffff;
+
+				&:disabled {
+					/* Appearance */
+					cursor: not-allowed;
+					opacity: .4;
+				}
+			}
+
+			&.widget_secondary-button {
+				/* Layout */
+				margin-inline-end: 3--step;
+
+				/* Appearance */
+				color: #005a8f;
+			}
+		}
+
+		& .widget_rating-group-wrapper {
+			display: flex;
+
+			/* Layout */
+			align-items: center;
+		}
+
+		& .current-url {
+			display: none;
+		}
+
+		& textarea, input {
+			/* Layout */
+			box-sizing: border-box;
+
+			/* Layout */
+			inline-size: 100%;
+
+			/* Layout */
+			padding: 2--step;
+
+			/* Text */
+			font: inherit;
+
+			/* Appearance */
+			background-color: #fafafa;
+			border: 1px solid #2f7aa7;
+
+			/* Appearance */
+			border-radius: 3px;
+		}
+
+		& textarea {
+			/* Layout */
+			block-size: 31--step;
+
+			/* Layout */
+			margin-block-start: 6--step;
+
+			&:focus {
+				/* Appearance */
+				outline: 1px solid #2f7aa7;
+			}
+		}
+
+		& input {
+			/* Layout */
+			margin-block-start: 3--step;
+
+			&:focus {
+				&:valid {
+					/* Appearance */
+					outline: 1px solid #2f7aa7;
+				}
+			}
+
+			&:invalid {
+				/* Appearance */
+				outline: 1px solid #ff3838;
+			}
+
+			&[type="radio"] {
+				display: none;
+			}
+		}
+
+		& .widget_rate-group {
+			display: flex;
+
+			/* Layout */
+			justify-content: center;
+			padding-inline-start: 8--step;
+		}
+
+		& .widget_button-group {
+			display: flex;
+
+			/* Layout */
+			justify-content: flex-end;
+		}
+
+		& .widget_success {
+			display: flex;
+			display: none;
+
+			/* Layout */
+			align-items: center;
+			block-size: 100%;
+			block-size: 0;
+			flex-direction: column;
+			inline-size: 100%;
+
+			/* Layout */
+			inset-block-start: 50--rpx;
+			inset-inline-end: 0;
+			justify-content: center;
+			position: absolute;
+
+			/* Appearance */
+			background-color: #ffffff;
+			opacity: 0;
+
+			/* Animation */
+			transition: opacity 200ms ease-in;
+
+			@media (width < 1024px) {
+				/* Layout */
+				inset-block-start: 0;
+			}
+
+			& .widget_success-interior {
+				/* Layout */
+				position: relative;
+			}
+
+			&.-active {
+				display: flex;
+
+				/* Layout */
+				block-size: 100%;
+
+				/* Appearance */
+				opacity: 1;
+			}
+
+			& svg {
+				fill: #005a8f;
+
+				&.success {
+					fill: #f68e1e;
+
+					/* Animation */
+					transition: fill 400ms ease-in;
+				}
+			}
+
+			& .widget_success-orange-circle {
+				display: flex;
+
+				/* Layout */
+				align-items: center;
+
+				/* Layout */
+				block-size: calc(5--step - 1px);
+				box-sizing: content-box;
+				inline-size: calc(5--step - 1px);
+				justify-content: center;
+				position: absolute;
+
+				& span {
+					/* Appearance */
+					display: block;
+
+					/* Layout */
+					block-size: 2--step;
+
+					/* Layout */
+					box-sizing: content-box;
+					inline-size: 2--step;
+
+					/* Appearance */
+					background-color: #005a8f;
+					border-radius: 100%;
+					opacity: 1;
+
+					/* Animation */
+					animation-duration: 1200ms;
+
+					/* Animation */
+					animation-iteration-count: infinite;
+					animation-name: antennaDot;
+					//transition: opacity 200ms ease-in;
+				}
+
+				&.circle-1 {
+					/* Layout */
+					/* inset-block-start: 146px;
+					inset-inline-start: 174px; */
+					inset-block-start: 26px;
+					inset-inline-start: 92px;
+
+					& span {
+						/* Layout */
+						display: block;
+
+						/* Animation */
+						animation-delay: 0s;
+					}
+				}
+
+				&.circle-2 {
+					/* Layout */
+					/* inset-block-start: 158px;
+					inset-inline-start: 161px; */
+					inset-block-start: 37px;
+					inset-inline-start: 78px;
+
+					& span {
+						/* Animation */
+						animation-delay: 400ms;
+					}
+				}
+
+				&.circle-3 {
+					/* Layout */
+					/* inset-block-start: 170px;
+					inset-inline-start: 148px; */
+					inset-block-start: 49px;
+					inset-inline-start: 63px;
+
+					& span {
+						/* Animation */
+						animation-delay: 800ms;
+					}
+				}
+			}
+
+			& .widget_success-text {
+				/* Layout */
+				margin-block-start: 0;
+
+				/* Layout */
+				padding-block-start: 10--step;
+
+				/* Text */
+				font-size: 24--rpx;
+				line-height: 28--rpx;
+				text-align: center;
+
+				/* Appearance */
+				color: #292a2d;
+			}
+		}
+
+		& .widget_fail {
+			display: flex;
+			display: none;
+
+			/* Layout */
+			align-items: center;
+			block-size: 100%;
+			block-size: 0;
+			flex-direction: column;
+			inline-size: 100%;
+
+			/* Layout */
+			inset-block-start: 0;
+			inset-inline-end: 0;
+			justify-content: center;
+			position: absolute;
+
+			/* Appearance */
+			background-color: #eaeef4;
+			opacity: 0;
+
+			/* Animation */
+			transition: opacity 200ms ease-in;
+
+			&.-active {
+				display: flex;
+
+				/* Layout */
+				block-size: 100%;
+
+				/* Appearance */
+				opacity: 1;
+			}
+
+			& svg path {
+				fill: #c8102e;
+			}
+
+			& .widget_fail-text {
+				/* Layout */
+				padding-block-start: 10--step;
+
+				/* Text */
+				font-size: 24--rpx;
+				line-height: 28--rpx;
+				text-align: center;
+
+				/* Appearance */
+				color: #292a2d;
+			}
+		}
+	}
+}
+
+/* Context Styles */
+
+:host([collapsible="true"]) {
+	& .widget_content {
+		& .widget_success {
+			/* Layout */
+			inset-block-start: 0;
+
+			/* Appearance */
+			background-color: #eaeef4;
+		}
+	}
+
+	& .widget_rating-group-wrapper {
+		/* Layout */
+		flex-direction: column;
+
+		& .widget_rate-group {
+			/* Layout */
+			padding-inline-start: 0;
+		}
+	}
+
+	& slot[name="title"] {
+		@media (width < 1024px) {
+			display: block;
+		}
+	}
+}
+
+@keyframes antennaDot {
+	from {
+		/* Appearance */
+		background-color: #005a8f;
+	}
+
+	30% {
+		/* Appearance */
+		background-color: #005a8f;
+
+		/* Transform */
+		transform: scale(1);
+	}
+
+	50% {
+		/* Appearance */
+		background-color: #f68e1e;
+
+		/* Transform */
+		transform: scale(1.5);
+	}
+
+	to {
+		/* Appearance */
+		background-color: #005a8f;
+	}
+}
