+++ src/components/feedback-widgets/feedback-widget-static/feedback-widget-static.css
@@ -15,382 +15,4 @@
 		/* Layout */
 		margin-inline-end: 0;
 	}
-
-	& .widget-static_content {
-		/* Layout */
-		position: relative;
-
-		/* Appearance */
-		color: #292a2d;
-
-		& h3 {
-			/* Text */
-			margin-block: 5--step;
-
-			/* Text */
-			font-size: 24--rpx;
-			font-weight: 400;
-			line-height: calc(20 / 18);
-		}
-
-		& p {
-			/* Text */
-			font-size: 16--rpx;
-		}
-
-		& button {
-			/* Layout */
-			margin-block-start: 6--step;
-
-			/* Layout */
-			padding-block: 1--step;
-			padding-inline: 4--step;
-
-			/* Appearance */
-			border: 1px solid #005a8f;
-
-			/* Appearance */
-			border-radius: 3px;
-			cursor: pointer;
-
-			&.widget-static_vote-button {
-				/* Layout */
-				margin-block-start: 0;
-				padding: 2--step;
-
-				&:first-child {
-					/* Layout */
-					margin-inline-end: 2--step;
-				}
-
-				&.selected {
-					/* Appearance */
-					background-color: #2f7aa7;
-					border: 1px solid #005a8f;
-
-					& svg path {
-						fill: #ffffff;
-					}
-				}
-			}
-
-			&.widget-static_primary-button {
-				/* Appearance */
-				background-color: #005a8f;
-				color: #ffffff;
-
-				&:disabled {
-					/* Appearance */
-					cursor: not-allowed;
-					opacity: .4;
-				}
-			}
-
-			&.widget-static_secondary-button {
-				/* Layout */
-				margin-inline-end: 3--step;
-
-				/* Appearance */
-				color: #005a8f;
-			}
-		}
-
-		& .current-url {
-			display: none;
-		}
-
-		& textarea, input {
-			/* Layout */
-			inline-size: 100%;
-
-			/* Layout */
-			padding: 2--step;
-
-			/* Text */
-			font: inherit;
-
-			/* Appearance */
-			background-color: #fafafa;
-			border: 1px solid #2f7aa7;
-
-			/* Appearance */
-			border-radius: 3px;
-		}
-
-		& textarea {
-			/* Layout */
-			block-size: 31--step;
-
-			/* Layout */
-			margin-block-start: 6--step;
-
-			&:focus {
-				/* Appearance */
-				outline: 1px solid #2f7aa7;
-			}
-		}
-
-		& input {
-			/* Layout */
-			margin-block-start: 3--step;
-
-			&:focus {
-				&:valid {
-					/* Appearance */
-					outline: 1px solid #2f7aa7;
-				}
-			}
-
-			&:invalid {
-				/* Appearance */
-				outline: 1px solid #ff3838;
-			}
-
-			&[type="radio"] {
-				display: none;
-			}
-		}
-
-		& .widget-static_rating-group-wrapper {
-			display: flex;
-
-			/* Layout */
-			align-items: center;
-
-			& p {
-				/* Layout */
-				margin-block-start: 0;
-			}
-
-			& .widget-static_rate-group {
-				display: flex;
-
-				/* Layout */
-				justify-content: center;
-				padding-inline-start: 8--step;
-			}
-		}
-
-		& .widget-static_button-group {
-			display: flex;
-
-			/* Layout */
-			justify-content: flex-end;
-		}
-
-		& .widget-static_success {
-			display: flex;
-			display: none;
-
-			/* Layout */
-			align-items: center;
-			block-size: 100%;
-			block-size: 0;
-			flex-direction: column;
-			inline-size: 100%;
-
-			/* Layout */
-			inset-block-start: 50--rpx;
-			justify-content: center;
-			position: absolute;
-
-			/* Appearance */
-			background-color: #ffffff;
-
-			/* Appearance */
-			opacity: 0;
-
-			/* Animation */
-			transition: opacity 200ms ease-in;
-
-			&.-active {
-				display: flex;
-
-				/* Layout */
-				block-size: 100%;
-
-				/* Appearance */
-				opacity: 1;
-			}
-
-			& svg {
-				fill: #005a8f;
-
-				&.success {
-					fill: #f68e1e;
-
-					/* Animation */
-					transition: fill 400ms ease-in;
-				}
-			}
-
-			& .widget-static_success-orange-circle {
-				display: flex;
-
-				/* Layout */
-				align-items: center;
-
-				/* Layout */
-				block-size: calc(5--step - 1px);
-				box-sizing: content-box;
-				inline-size: calc(5--step - 1px);
-				justify-content: center;
-				position: absolute;
-
-				& span {
-					/* Appearance */
-					display: block;
-
-					/* Layout */
-					block-size: 2--step;
-
-					/* Layout */
-					box-sizing: content-box;
-					inline-size: 2--step;
-
-					/* Appearance */
-					background-color: #005a8f;
-					border-radius: 100%;
-					opacity: 1;
-
-					/* Animation */
-					animation-duration: 1200ms;
-
-					/* Animation */
-					animation-iteration-count: infinite;
-					animation-name: antennaDot;
-					//transition: opacity 200ms ease-in;
-				}
-
-				&.circle-1 {
-					/* Layout */
-					inset-block-start: 109px;
-					inset-inline-end: 304px;
-
-					& span {
-						/* Layout */
-						display: block;
-
-						/* Animation */
-						animation-delay: 0s;
-					}
-				}
-
-				&.circle-2 {
-					/* Layout */
-					inset-block-start: 121px;
-					inset-inline-end: 315px;
-
-					& span {
-						/* Animation */
-						animation-delay: 400ms;
-					}
-				}
-
-				&.circle-3 {
-					/* Layout */
-					inset-block-start: 134px;
-					inset-inline-end: 328px;
-
-					& span {
-						/* Animation */
-						animation-delay: 800ms;
-					}
-				}
-			}
-
-			& .widget-static_success-text {
-				/* Layout */
-				margin-block-start: 0;
-
-				/* Layout */
-				padding-block-start: 10--step;
-
-				/* Text */
-				font-size: 24--rpx;
-				line-height: 28--rpx;
-				text-align: center;
-
-				/* Appearance */
-				color: #292a2d;
-			}
-		}
-
-		& .widget-static_fail {
-			display: flex;
-			display: none;
-
-			/* Layout */
-			align-items: center;
-			block-size: 100%;
-			block-size: 0;
-			flex-direction: column;
-			inline-size: 100%;
-
-			/* Layout */
-			inset-block-start: 50--rpx;
-			justify-content: center;
-			position: absolute;
-
-			/* Appearance */
-			background-color: #ffffff;
-
-			/* Appearance */
-			opacity: 0;
-
-			&.-active {
-				display: flex;
-
-				/* Layout */
-				block-size: 100%;
-
-				/* Appearance */
-				opacity: 1;
-			}
-
-			& .widget-static_fail-text {
-				/* Layout */
-				margin-block-start: 0;
-
-				/* Layout */
-				padding-block-start: 10--step;
-
-				/* Text */
-				font-size: 24--rpx;
-				line-height: 28--rpx;
-				text-align: center;
-
-				/* Appearance */
-				color: #292a2d;
-			}
-		}
-	}
-}
-
-@keyframes antennaDot {
-	from {
-		/* Appearance */
-		background-color: #005a8f;
-	}
-
-	30% {
-		/* Appearance */
-		background-color: #005a8f;
-
-		/* Transform */
-		transform: scale(1);
-	}
-
-	50% {
-		/* Appearance */
-		background-color: #f68e1e;
-
-		/* Transform */
-		transform: scale(1.5);
-	}
-
-	to {
-		/* Appearance */
-		background-color: #005a8f;
-	}
 }
